move Raster, etc. into schemas
